+++ pages/_app.js
@@ -1,4 +1,4 @@
-import { PageLayout } from "@/layouts/PageLayout";
+// import { PageLayout } from "@/layouts/PageLayout";
 import { QueryClient, QueryClientProvider } from "react-query";
 import "../styles/globals.css";
 
@@ -7,9 +7,8 @@ const queryClient = new QueryClient();
 function MyApp({ Component, pageProps }) {
   return (
     <QueryClientProvider client={queryClient}>
-      <PageLayout>
-        <Component {...pageProps} />
-      </PageLayout>
+      {/* <PageLayout></PageLayout> */}
+      <Component {...pageProps} />
     </QueryClientProvider>
   );
 }
