--- styles/Home.module.css
@@ -1,41 +1,0 @@
-.Container {
-  display: grid;
-  grid-template-columns: 1fr min(65ch, 100%) 1fr;
-}
-
-.Container > * {
-  grid-column: 2;
-}
-
-.Full {
-  width: 100%;
-  grid-column: 1/4;
-}
-
-.LinkTree {
-  margin-top: 24px;
-}
-
-.LinkTree .Section {
-  margin-bottom: 64px;
-  padding: 0 8px;
-}
-
-.LinkTree .Headings {
-  background: white;
-  color: black;
-  padding: 2px 14px;
-  margin-bottom: 16px;
-  display: flex;
-  width: 100%;
-  justify-content: center;
-}
-
-.LinkTree .link {
-  display: block;
-  margin-bottom: 24px;
-  text-align: center;
-  font-family: var(--bebas);
-  border: 1px solid white;
-  padding: 4px;
-}
